+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -7,6 +7,8 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import roomescape.controller.interceptor.CheckRoleInterceptor;
+import roomescape.controller.resolver.LoginUserArgumentResolver;
 import roomescape.service.AuthService;
 import roomescape.service.UserService;
