+++ src/main/java/roomescape/controller/interceptor/CheckRoleInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.controller.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
